refactor: Move multi-session components and update imports
Reorganize multi-session UI by moving several components into components/multi-session (manage-account(s), switch-account-item, switch-account-menu, switch-account-menu-item) and update relative imports accordingly (UserView and multiSessionPlugin paths). Update multi-session plugin to import ManageAccounts and SwitchAccountMenuItem from the new location, adjust top-level exports (remove old settings/account exports and re-export multi-session manage-accounts from plugins.ts). This refactor groups multi-session features and fixes broken import/export paths.
