+++ packages/heroui/src/components/multi-session/switch-account-item.tsx
@@ -5,7 +5,7 @@ import {
 } from "@better-auth-ui/react"
 import { Dropdown, Spinner } from "@heroui/react"
 import type { Session, User } from "better-auth"
-import { UserView } from "./user-view"
+import { UserView } from "../user/user-view"
 
 type DeviceSession = {
   session: Session
