+++ packages/heroui/src/index.tsx
@@ -19,8 +19,6 @@ export * from "./components/email/reset-password-email"
 export * from "./components/settings/account/account-settings"
 export * from "./components/settings/account/appearance"
 export * from "./components/settings/account/change-email"
-export * from "./components/settings/account/manage-account"
-export * from "./components/settings/account/manage-accounts"
 export * from "./components/settings/account/user-profile"
 export * from "./components/settings/security/active-sessions"
 export * from "./components/settings/security/change-password"
