+++ packages/heroui/src/lib/multi-session/multi-session-plugin.ts
@@ -3,8 +3,8 @@ import {
   type MultiSessionPluginOptions
 } from "@better-auth-ui/core/plugins"
 
-import { ManageAccounts } from "../../components/settings/account/manage-accounts"
-import { SwitchAccountMenuItem } from "../../components/user/switch-account-menu-item"
+import { ManageAccounts } from "../../components/multi-session/manage-accounts"
+import { SwitchAccountMenuItem } from "../../components/multi-session/switch-account-menu-item"
 import type { AuthPlugin } from "../auth-plugin"
 
 export function multiSessionPlugin(options: MultiSessionPluginOptions = {}) {
