+++ packages/heroui/src/plugins.ts
@@ -3,7 +3,7 @@
 // Plugin-contributed components that may be used standalone
 export * from "./components/auth/magic-link"
 export * from "./components/passkey/passkeys"
-export * from "./components/settings/account/manage-accounts"
+export * from "./components/multi-session/manage-accounts"
 export * from "./lib/auth-plugin"
 export * from "./lib/magic-link/magic-link-plugin"
 export * from "./lib/multi-session/multi-session-plugin"
